Refactor into System.Win32.Compat
After #55, the CPP pragmas in module `System.Console.ANSI.Windows.Foreign` are there to provide functions available in later versions of the `Win32` package.

This proposal refactors the code to make that explicit. The intention being to make the logic more explicit.
